+++ CHANGELOG.rst
@@ -1,6 +1,12 @@
 Locks library change log
 ========================
 
+2.2.0
+-----
+
+  * ADDED: Tests now run on xcore.ai as well as xcore-200
+  * ADDED: Support for XCommon CMake build system
+
 2.1.0
 -----